refactor(video-editor): extract overlay renderer from canvas (#4513)
Move the letterbox/scrim rendering (CanvasFitter, VideoSetupLoadingIndicator,
overlay cut area, scrim bars), the HitTestExpander render object, and the
scrimZoomTransform helper out of video_editor_canvas.dart into dedicated
sibling modules, so editor work no longer serializes on one 2.4k-line file.

Pure code move, no behavior change: canvas drops 2448 -> 2068 lines. The two
concern-named tests now import their real modules. Drops the now-inaccurate
@visibleForTesting on HitTestExpander/scrimZoomTransform, which are used from
production code after the split.

First, low-risk slice of #4513; the <500-line target needs extracting the
stateful player/gesture/timeline logic from _VideoEditorState, deferred.
