+++ mobile/test/widgets/video_editor/main_editor/hit_test_expander_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:openvine/widgets/video_editor/main_editor/video_editor_canvas.dart';
+import 'package:openvine/widgets/video_editor/main_editor/hit_test_expander.dart';
 
 void main() {
   group(HitTestExpander, () {
