+++ mobile/test/widgets/video_editor/main_editor/scrim_zoom_transform_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:openvine/widgets/video_editor/main_editor/video_editor_canvas.dart';
+import 'package:openvine/widgets/video_editor/main_editor/scrim_zoom_transform.dart';
 
 /// Builds a scale+translate editor matrix the way the editor's
 /// `onEditorZoomMatrix4Change` reports pinch-zoom (no rotation/skew).
